api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
PyMySQL==1.1.1
psycopg[binary]==3.2.10
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
pydantic-settings==2.10.1
email-validator==2.3.0
jinja2==3.1.6
python-multipart==0.0.20
requests==2.32.5
argon2-cffi==25.1.0
thefuzz==0.22.1
python-Levenshtein==0.27.1
openai==1.107.2
